refactor(Prompts): update main exports and external imports
