--- client/src/components/Prompts/components/index.ts
@@ -1,1 +1,0 @@
-export { default as CreatePromptButton } from './CreatePromptButton';
