+++ client/src/components/Prompts/index.ts
@@ -1,10 +1,51 @@
-export { default as PromptName } from './PromptName';
-export { default as PromptsView } from './PromptsView';
-export { default as PromptEditor } from './PromptEditor';
-export { default as PromptForm } from './PromptForm';
-export { default as PreviewLabels } from './PreviewLabels';
-export { default as PromptGroupsList } from './Groups/List';
-export { default as DashGroupItem } from './Groups/DashGroupItem';
-export { default as EmptyPromptPreview } from './EmptyPromptPreview';
-export { default as PromptSidePanel } from './SidePanel/GroupSidePanel';
-export { default as CreatePromptForm } from './Groups/CreatePromptForm';
+// Layouts
+export { PromptsView } from './layouts';
+
+// Forms
+export { PromptForm, CreatePromptForm, VariableForm, PromptLabelsForm } from './forms';
+
+// Dialogs
+export { PreviewPrompt, DeleteVersion, VariableDialog, SharePrompt } from './dialogs';
+
+// Editor
+export { PromptEditor, VariablesDropdown, CodeVariableGfm, PromptVariableGfm } from './editor';
+
+// Fields
+export { PromptName, Command, Description, CategorySelector } from './fields';
+
+// Display
+export { PromptDetails, PromptVariables, PromptVersions, EmptyPromptPreview } from './display';
+
+// Buttons
+export {
+  CreatePromptButton,
+  AdminSettings,
+  AdvancedSwitch,
+  AlwaysMakeProd,
+  AutoSendPrompt,
+  BackToChat,
+  ManagePrompts,
+} from './buttons';
+
+// Lists
+export {
+  List as PromptGroupsList,
+  DashGroupItem,
+  ChatGroupItem,
+  ListCard,
+  NoPromptGroup,
+} from './lists';
+
+// Sidebar
+export {
+  GroupSidePanel as PromptSidePanel,
+  PromptsAccordion,
+  FilterPrompts,
+  PanelNavigation,
+} from './sidebar';
+
+// Utils
+export { CategoryIcon, SkeletonForm } from './utils';
+
+// Backwards compatibility - old export name
+export { PromptLabelsForm as PreviewLabels } from './forms';
