+++ client/src/locales/en/translation.json
@@ -632,6 +632,7 @@
   "com_ui_2fa_verified": "Successfully verified Two-Factor Authentication",
   "com_ui_accept": "I accept",
   "com_ui_action_button": "Action Button",
+  "com_ui_actions": "Actions",
   "com_ui_active": "Active",
   "com_ui_add": "Add",
   "com_ui_add_code_interpreter_api_key": "Add Code Interpreter API Key",
@@ -1053,9 +1054,11 @@
   "com_ui_link_copied": "Link copied",
   "com_ui_link_refreshed": "Link refreshed",
   "com_ui_loading": "Loading...",
+  "com_ui_labels": "Labels",
   "com_ui_locked": "Locked",
   "com_ui_logo": "{{0}} Logo",
   "com_ui_low": "Low",
+  "com_ui_make_production": "Make Production",
   "com_ui_manage": "Manage",
   "com_ui_marketplace": "Marketplace",
   "com_ui_marketplace_allow_use": "Allow using Marketplace",
@@ -1150,6 +1153,7 @@
   "com_ui_no_category": "No category",
   "com_ui_no_changes": "No changes were made",
   "com_ui_no_individual_access": "No individual users or groups have access to this agent",
+  "com_ui_no_labels": "No Labels",
   "com_ui_no_mcp_servers": "No MCP servers yet",
   "com_ui_no_mcp_servers_match": "No MCP servers match your filter",
   "com_ui_no_memories": "No memories. Create them manually or prompt the AI to remember something",
@@ -1160,6 +1164,7 @@
   "com_ui_no_read_access": "You don't have permission to view memories",
   "com_ui_no_results_found": "No results found",
   "com_ui_no_terms_content": "No terms and conditions content to display",
+  "com_ui_no_versions": "No versions",
   "com_ui_none": "None",
   "com_ui_not_used": "Not Used",
   "com_ui_nothing_found": "Nothing found",
@@ -1198,6 +1203,7 @@
   "com_ui_preview": "Preview",
   "com_ui_privacy_policy": "Privacy policy",
   "com_ui_privacy_policy_url": "Privacy Policy URL",
+  "com_ui_production": "Production",
   "com_ui_prompt": "Prompt",
   "com_ui_prompt_group_button": "{{name}} prompt, {{category}} category",
   "com_ui_prompt_group_button_no_category": "{{name}} prompt",
@@ -1215,6 +1221,7 @@
   "com_ui_prompts_allow_share": "Allow sharing Prompts",
   "com_ui_prompts_allow_share_public": "Allow sharing Prompts publicly",
   "com_ui_prompts_allow_use": "Allow using Prompts",
+  "com_ui_prompt_delete_confirm": "Are you sure you want to delete the '{{0}}' prompt?",
   "com_ui_provider": "Provider",
   "com_ui_quality": "Quality",
   "com_ui_read_aloud": "Read aloud",
